Calculation without entities
